moving `save_npy` stuff out of `triangulate_3d`
