+++ freemocap/data_layer/recording_models/recording_info_model.py
@@ -11,7 +11,7 @@ from freemocap.system.paths_and_filenames.file_and_folder_names import (
     OUTPUT_DATA_FOLDER_NAME,
     RAW_DATA_FOLDER_NAME,
     TOTAL_BODY_CENTER_OF_MASS_NPY_FILE_NAME,
-    MEDIAPIPE_REPROJECTION_ERROR_NPY_FILE_NAME,
+    RAW_MEDIAPIPE_REPROJECTION_ERROR_NPY_FILE_NAME,
     SYNCHRONIZED_VIDEOS_FOLDER_NAME,
     ANNOTATED_VIDEOS_FOLDER_NAME,
     MEDIAPIPE_3D_NPY_FILE_NAME,
@@ -102,7 +102,7 @@ class RecordingInfoModel:
             Path(self._path)
             / OUTPUT_DATA_FOLDER_NAME
             / RAW_DATA_FOLDER_NAME
-            / MEDIAPIPE_REPROJECTION_ERROR_NPY_FILE_NAME
+            / RAW_MEDIAPIPE_REPROJECTION_ERROR_NPY_FILE_NAME
         )
 
     @property
